+++ plugins/mbb-ppt-generator/skills/mbb-ppt-generator/mbb_ppt/__init__.py
@@ -22,4 +22,4 @@ from .review import SlideReviewer, AutoFixPipeline, review, autofix
 
 ExecEngine = MbbEngine
 
-__version__ = '0.5.1'
+__version__ = '0.5.2'